Move classes to better locations
