--- detekt-api/src/main/kotlin/io/gitlab/arturbosch/detekt/api/dsl/Config.kts
@@ -1,11 +1,0 @@
-package io.gitlab.arturbosch.detekt.api.dsl
-
-rules {
-    ruleSet("comments") {
-        active = false
-        undocumentedPublicClass {
-            active = true
-            searchInInnerClass = false
-        }
-    }
-}
